Don't cheat in memget anymore.
We used to materialize a fake PGAS address that worked for get emulation, but
this doesn't work in AGAS because fake addresses are not mapped in the
btt. Actually send along the return address this time. This implementation uses
internally knowledge to implement a special call_cc for this case to avoid a
copy that would be required by hpx_call_cc.


Former-commit-id: 6cd4c1fae326face26cf217b254182b405eb331a