This giant commit re-implements the entire address space management infrastructure to use the new jemalloc allocator interface.
Jemalloc now has the ability to explicitly designate thread caches for
allocation. Previously, we had to use multiple instances of the jemalloc library
in order to make sure that allocations didn't wind up cached in the wrong
place. We no longer need to worry about this impact. Furthermore, we previously
had to constrain jemalloc's ability to purge data where we can now explicitly
handle that event.

Overall, this has substantially reduced the complexity of our allocation
code. There are some opportunities remaining for simplification in the PGAS
infrastructure, and AGAS has not yet been updated.

This results in no obvious performance difference for the distributed memory
case. SMP execution of fibonacci is about 2x slower with --enable-jemalloc. It's
not clear why that is, if it's a performance regression in jemalloc or if it has
to do with new configuration details (e.g., their default chunk size dropped
from 4MB to 256k, which may be problematic for us).


Former-commit-id: 32bca6da71bede23eb6b87cd6d605f396cbdb632