Updated an assert message in the LCO implementation.
We have found that, when an LCO's vtable pointer is NULL, then nine times out of
ten it means that there is a use-after-free of an LCO. The NULL arises because
we zero-out the vtable pointer during delete specifically to catch this
problem. I updated the assert message to indicate that the use-after-free is
probably the issue.

I also made a couple of other asserts in this file more specific.


Former-commit-id: 205c97013fbe7dce53fd02f30ec8658bac8b2adf