Consider "0" count for gencounts as implicitly set.
This normalizes the view of the counter on the set side and the wait
size. Previously, the waiter would look one generation ahead of the setter. This
sort of worked, but we would get unexpected wakeups that would cause problems.

Now, the generation 0 is set at allocation, i.e., anyone waiting for generation
0 will not wait. The first set operation will set generation 1.


Former-commit-id: de621f811c98dd305b47489c03998569285cdad7