Remove the "core" variables from the scheduler and worker.
Now that we rely on a launcher we don't care about cores directly. The affinity
mask for the worker is set externally, so we just go ahead and spawn pthreads
and expect the system to deal with everything for us.

Going forward, we may want more control over the affinity of pthreads, but we
will use a more complicated hwloc-based mapping so this core information is
completely redundant.


Former-commit-id: 7d4a611ee682d7ddb9c5498ce0d462b1f2d1b813