Create an option for the user to control the stack cache limit.
Now that we cache stacks in freelists we need a way to control how many we let
accumulate. I added a runtime option to control the limit. -1 means no limit, 0
means no caching, and the default is 32.

In order to try and get some temporal locality from the allocator, we return
half of the stacks when we hit the limit.


Former-commit-id: b45a9ff2ffe16f3dabca18c22582330af4e85dd1