Fixing headers for instrumentation tracing
Our "named values" (i.e. constants) being written to the header were
being written as raw C structs. Turns out it was necessary to pack them
so that the reader could handle them properly. (We could have changed it
on either end; this is just the way it worked out.)


Former-commit-id: c1bc946cefe43f925c3aefbe6b23057cfdb41bc9