Use a #define instead of a static const int.
For whatever reason Clang emits a warning about some gnu extension when we use a
static const int as the size of a dynamically allocated array. I switched to
using a #define instead of the constant in order to avoid that.

We could also suppress that warning just for Clang, but it's easier to just use
the #define. This is just a simple test case.


Former-commit-id: 27c3da9a85cbd7ddb78767be0f6a47b84bff5dd5