AGAS: Make hpx_gas_free block until there are outstanding pins.
In AGAS, hpx_gas_free removed the btt entry directly without checking
the pinned reference count of the block. This resulted in a scenario (in the
lco_collectives test, where an LCO is set from a pinned action) where we would
try to delete a block that was not entirely unpinned.

This change makes the gas-free operation block until the pinned reference
count for each block in the allocation reaches 0.


Former-commit-id: 0741801621b6a640ca238d1e84af9ffe6e20ae3c