Stop doing custom aligned mmap() for the PGAS heap.
We have added a system interface that provides an aligned interface to mmap, so
we don't need the custom heap allocation code anymore. The one potential issue
is that the system mmap() overallocates virtual address space and then trims
unneeded bits of the allocation with munmap. This can be an issue where
overcommit isn't available.


Former-commit-id: 211e756c8b260ce4dcc9dd426a9ec5515f061fa0