Suggest a properly aligned address for heap allocation.
The basic protocol for aligned mmap() operations is to first try to mmap() the
required number of bytes, hoping that the alignment works out, and mmap()ing a
larger region and trimming the excess with munmap() if necessary.

The global heap tends to be a large allocation with large alignment
requirements, and it appears to generally take the second code path due in large
part to the alignment requirement. The second code path is more expensive and
can cause problems when overcommit is unavailable however.

This commit tries to help the lucky mmap() code path by suggesting a base
address that has the correct alignment. This is what the older heap-specific
mmap() code used to do.


Former-commit-id: 3e17839215ec00383bc942f4dad9edbb64472097