Export the libhpx configuration so the user can query it programatically.
This allows an application to query the current configuration. It is done in a
brittle way, installing the libhpx headers related to configuration and exposing
the global pointer to here->config. This structure can vary between versions.

It would be more portable to write a string-query interface, which we may want
to do at some point.


Former-commit-id: 76ae327460733de9b0d6529a380196c6ee710545