Fix an issue with freeing array allocations in AGAS.
While freeing array allocations (blocks greater than 1) in GAS, we
would prematurely give global memory back to jemalloc and signal the
user that a free is done. This could lead to a case where we would
give invalid memory back.

This commit fixes the above issue by using an "AND" lco to wait until
all of the blocks in an allocation have been unpinned and freed.


Former-commit-id: 4455b370c4b9637777cc6e705e5107071a475551