Don't complain about ranks=1 in PGAS initialization.
We used to perform a check in the initializer and create an error when PGAS is
initialized if the number of ranks was equal to 1. In fact, PGAS works fine with
one rank, and we would like to be able to run single rank PGAS for correctness
and performance debugging reasons.


Former-commit-id: b3692aad981fb0fea61e256495178ccfcfa672f4