Add an implementation of remote lco_get for the IS/IR network.
The Isend/Irecv network uses parcels and the scheduler_wait_launch_when()
operation to implement an optimized synchronous hpx_lco_get operation for remote
LCOs. This allows us to bypass the allocation of a local proxy future for the
operation, along with the memcpy operation to it.

Unfortunately, the structure of the reply handler currently requires us to
perform an extra copy of the data on the remote side, however parcel
environments and continuation stacks, or explicit access to the continuation
parcel, can eliminate that copy in the future.


Former-commit-id: 57c466f18a420a0b524b4efd20f94e12aa7e7d38