Replace scheduler_wait_send_through with a generic scheduler_suspend.
This functionality is quite convenient. The basic situation we are dealing with
is that a lightweight thread wants to suspend want wait for some event (other
than an LCO signal) to cause it to resume, and this event is dependent on
something it will do.

As an example, the thread night want to wait for a synchronous memget to
complete. We would prefer not to allocate a proxy LCO for this operation, since
that is global memory churn and locking. Instead, we want to tie a resume
command directly to the lsync operation passed to memget.

The problem is that, if we issue the memget from the thread that will wait, the
event (lsync command on the memget) happens asynchronously. In particular, it
could happen instantaneously from the perspective of the issuing thread. This
would trigger the resume, and would reschedule the thread somewhere to
run. Unfortunately, the thread 1) has not yet checkpointed itself and 2) is
actually still running. Executing the same thread concurrently will be
bad. Other races exist as well.

This patch provides scheduler_suspend(), which allows the calling thread to run
code *after* it has been checkpointed and suspended, but before a new thread
runs. In the memget example, it can issue the memget from that context and be
sure there is no wakeup race.

I expect this to be extremely useful in many cases. It is more generic than
scheduler_wait() which could be implemented by scheduler_resume() but is common
enough to specialize for, and could also be used for more intelligent
scheduler_yield() experimentation.


Former-commit-id: f7aa5ce1a486d3439ff6be27731a82975c9de76f