fabric/man: Clarify when rx/tx attr mode & caps bits are valid
Mode bits are defined as part of fi_info, fi_rx_attr, and
fi_tx_attr.  There's confusion around what it means with
the rx/tx attr mode bits are set to 0, but fi_info mode
bits are not.

The intent is that the rx/tx attr mode bits be ignored on
input to fi_getinfo when they are 0, but fi_info mode is
not.  That is, fi_info:mode bits override the rx/tx attr
mode bits.

The same issue arises with the caps bits.

Document this intent in the man pages.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>