fabric/param: Use dlist for param list
Use the double linked list abstraction in place of a custom
linked list.  This will allow us to insert items at the tail
of the list, rather than the head, which keeps the core params
at the start of the list, rather than the end.  It will also
allow us to remove params that are deregistered.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>