The proxy library isn't necessary for GAS allocation with tbb.
The proxy library contains the symbol bindings for the symbols in malloc.h. For
the GAS allocator we are not using these bindings. At the same time, it seems
plausible that we would like to allow the user to use tbbmalloc for their
default allocator as well, without requiring them to LD_PRELOAD the library. We
will leave this a future work.


Former-commit-id: 2a344aacef4644755061486f13a675333a44b87a