Modify libffi contrib configuration.
We depend on libffi availability and we don't want to require libffi as an
external dependency, so we distribute it in our contrib directory and can
configure/build it with HPX. At the same time, libffi *is* available in many
systems, and some systems have custom versions. Given that we would prefer to
avoid the overhead of configure/build this patch will first look for a system
installed libffi and use that if available.


Former-commit-id: 1df054bae619f6351d069647a17343f44639cc93