Update the hwloc configuration.
We want to be able to depend on external installations of hwloc, either in our
path or through pkg-config. We don't want to have a custom hpx_ prefix in this
case, so I removed that from our contrib-ed configuration and the source code
that uses hwloc.

In addition, I updated the configuration file to allow the specification of
hwloc though an updated --with-hwloc option. If hwloc is found (either through
the system or through the contrib) then we assume that the user wants to use
hwloc functionaliy, so I removed the separate --enable-hwloc option.


Former-commit-id: 444a34c6bccc5994fddd7f1ce9004a26b84b6b3b