Add a worker_can_alloc() query that we can use to alloca() safely.
We tend to use relatively small stacks, so we need to be careful inside of
libhpx about dynamic stack allocation. At the same time, stack allocation is
efficient, particularly for temporaries.

Fortunately, we actually know what the bottom of the stack is, since we are in
control of the stacks. We can use this to determine the number of bytes left in
the stack to estimate if an alloca is going to be safe.


Former-commit-id: d3bb3417d7bd6bb1bd789e5084053b06b0bd6008