Stop using getref() in the pwc rendezvous put optimization.
The getref was causing a timing race where a remote get operation completes
before all of the buffer references created during the get operation have been
released. In this case, it is valid code to delete the associated LCO once the
get is complete, however this was causing use-after-free when the buffer
reference was released.

The fix is to use get() to the stack or to a malloced buffer. This operation
creates a temporary copy, but avoids the race condition. A third option exists,
perform the rendezvous put operation as a pure put() without a completion, then
release the buffer, then perform a second put() operation for the resume
command. This trades the temporary copy for two put() operations. This will be
faster depending on the size of the buffer. This algorithm is implemented but
commented out, as we don't have a great idea about when it is a benefit, and the
network_command() operation is not being received correctly.


Former-commit-id: 73327ae8e02e3756af2b4484db6ade2b37e3080c