Fix deallocation issues in AGAS.
In AGAS, for cyclic allocations we have to make sure that we free up all
of the blocks on non-root localities before we free the blocks that were
originally allocated on the root locality (since the root locality controls
the address space of the allocation).

These changes take into account that blocks can move and each free follows
the block. All of the cases (block moves from or to root locality) are
handled.


Former-commit-id: 5f1706bc26dec499f6e4e9cb22d982a8becad264