Handle the --hpx-trace-dir= runtime option differently.
If this attribute is unspecified, we use a generated directory in /tmp. If the
option is specified and is a (relative) path to a directory, we will output
trace files in the directory. If the option is specified but the path doesn't
exist we will create a directory at the path. If the option is specified with a
path that is not a directory then we disable tracing.


Former-commit-id: 5077ef957a8f624ff8efff31df6621774076ea05