Use usleep() instead of nop parcel spin in the schedule loop.
When the scheduler couldn't find work to do, we have always implemented
"spinning" through a nop parcel loop. In other words, we spin by instantiating a
thread that does nothing, and then calls schedule again.

This patch spins by calling usleep(1) instead. This reduces contention on
important data structures, which results in HPX "doing nothing, better." It has
no impact when stealing is likely to be successful (i.e., fibonacci). It will
introduce latency into detecting load imbalance (as it could take a microsecond
for us to detect changes in load), but this is likely to be fine given the
latencies of high-speed networks. We could probably increase 1 to 5 and still be
fine.

I abstracted usleep() as system_usleep() and also adjusted how we configure the
system directory as part of this patch as well.


Former-commit-id: e5fda8cca0696961ed4ba9bc6d70a848b8e4d3d2