Update the remote get handler to use all three handler implementations.
With the put-with-command network we have tried to optimize the remote-get-LCO
functionlity by turning it into as simple a rendezvous-put operation as
possible. This is unexpectedly complex.

We can't simply do the rendezvous-put-with-command to finish the remote get
operation. If we do, the higher-level get operation completes before the
rendezvous-put handler is done, and the underlying LCO can be deleted before the
put gets to cleanup, which results in Bad Things (TM).

In order to deal with this, we need to either 1) prevent the higher level get
operation from completing before the rendezvous put operation gets a chance to
cleanup or 2) perform the rendezvous-put operation in a way where the cleanup
operation isn't dependent on the underlying LCO buffer. A third option is
available, which is to delay a "LCO delete" operation from completing while the
LCO has a reference exposed---this requires reference counting and is something
we can do with the AGAS runtime, but not with the PGAS runtime.

This patch fully enables our solution.

1) If the LCO buffer is small enough that we don't mind making a temporary copy
   of it then we use the lco_get operation to copy the buffer out, before
   performing the rendezvous put. This solves the problem because the cleanup
   operation for this option doesn't depend on the LCO, so the dangerous
   ordering doesn't happen.

   a) If the buffer will fit on our stack, then we use the stack as the
      temporary buffer for the copy. This is optimal because we don't have to
      hit the allocator, and the stack is already registered for rdma.

   b) The the buffer doesn't fit on our stack, then we use registered_malloc()
      to get a buffer that we then registered_free() during cleanup.

2) If the buffer is big then we use lco_getref() to get a reference to its
   internal buffer. We then put this back to the requester, but don't send a
   completion command. After the first put completes, we can release the buffer
   and then send the completion command seperately. Technically, these commands
   aren't ordered and the completion could arrive before the buffer has been
   written. In fact, a put-without-remote-command is implemented as pure rdma,
   and the local completion indicates remote completion and proper ordering, so
   we're safe.


Former-commit-id: f326cd394f896e8eccdcf566f772553ee74827e1