Fix a race in the "and" LCO.
We used to use the and's value in the lco_get/wait handler to determine if we
should wait. When the value was 0 we would return. In the set handler we used
atomic fetch-and-add in order to manipulate the value, and when the value
reached 0 we would acquire the lock and signal all waiting threads.

This exposed a timing window between setting the value to 0 and acquiring the
lock. In that window we could see a rapid pair of wait+delete events on the and
gate sneak in. This is valid from the user level because set/wait/delete should
all serialize. Unfortunately, the delete gives the memory back to the allocator
and all sorts of things (including nothing) can happen to the LCO in that
window, before the setter gets to run the signal all code.

I fixed this problem by using the existing LCO "triggered" state to indicate
that the and LCO has been triggered. This state is only ever accessed while
holding the LCO lock, so the timing window that allows the wait+delete through
can no longer happen (the wait will block on the triggered state, not value ==
0).


Former-commit-id: 166a91226b18a7bfa0abf1442d3192b4156588ab