Forward --enable-static and --enable-shared to contrib projects.
For contrib projects that we configure using the ACX_CONFIGURE_DIR and that use
libtool (Photon and libffi for now), it makes sense to propagate the values for
these two flags. This patch means that a static build of hpx, when using the
contrib ffi or photon, will statically link those two libraries.


Former-commit-id: ea27db594b5de797b269dd8b1c378724f0e81220