Update the libffi to avoid a configure bug on cutter.
I found that the intel module on cutter doesn't seem to correctly add paths to
find ffi.h. I'm not sure that I understand why. The libffi package seems to
assume that its header was installed in a visible location. I was encountering
an issue where ffi.h couldn't be found. I added some verification code to make
sure that, after finding the package both ffi.h and libffi are available---if
they aren't we fall back to the built-in libffi.


Former-commit-id: 6b9292bdde6c17416f2d6f69f27f2ed093329d02