Partially implement a _CASM language for autoconf.
This means "preprocessed ASM" and in practice means that we can test for
features using .S files. This was necessary on Phi machines to test for the
presece of cmpxchg16b instructions. This *does not* fully support everything
related to a language, but does enough to give us what we need, which is engough
to do the following.

AC_PUSH_LANG([_CASM])
AC_COMPILE_IFELSE([AC_LANG_SOURCE([[ asm: in here; ]])], [], [])
AC_POP_LANG


Former-commit-id: 84be4345db6eec3797a57b8e53483cb43f4733bf