Clean up the worker infrastructure.
Make the 'self' pointer volatile, because it can change unexpectedly if a
lightweight thread is moved between workers. In order to reduce the number of
volatile load operations and TLS hits we extend the worker interface so that the
worker structure pointer always needs to be passed, and we pass around the
worker pointer as a parameter internal to the worker code wherever that makes
sense.

Don't store a backpointer from the worker struct to the scheduler, the scheduler
is always available as here->sched.

Also update some style nits, including adding a typedef for the worker struct.


Former-commit-id: 591e8b2896ea724df9c3a563b1f603bbc081f845