Suppress log output in IS/IR network progress when no messages complete.
This cuts way down on the amount of noise from the logger, in particular during
debugging when a thread is suspended in usleep(). It can suppress some
interesting information in that it suprresses the fact that progress is ongoing,
but the net log level is not useful when it is enabled.


Former-commit-id: 57cc3d48b45568c62937ab6c3627d731b8484303