Implement simplified AGAS free operation.
This may not be the most sophisticated free implementation, but it seems to work
(at least it passes the test suite) and is a good starting point for more
complicated algorithms.

The primary free functionality simply checks to see if it can "fastpath" the
free operation, which means free a single block that has no outstanding
references and has not been relocated. If any of these conditions is not met,
then it launches a basic hierarchical tree traversal.

for each segment in allocation
  for each block in segment
    block free
  free segment
free allocation


Former-commit-id: 5977170e70afcec32eba23fb6693bce8dfaf4b1a