Update the sendrecv example to use modern interface, and to be a lengthy test.
The sendrecv "example" was inspired by some MPI-ish code that attempts to
overlap inter-rank communication with computation. The example itself didn't
make much sense though, as it wasn't clear /why/ certain things were
happening. We decided that it has a role as a lengthy-test, as it seems to
trigger failures in corner cases, and it measures the cost of allocating big
parcels, so we moved it to the performance tests.


Former-commit-id: fb1d266eb0067771356709f94b6e053bd4a4f638