Modified the scheduler_suspend transfer type to take a parcel pointer.
After some experience with scheduler_suspend, it became obvious that suspend
continuations always pass the current parcel in their environment in some
way. This makes sense, since the continuation needs to do *something* that will
ultimately result in that parcel being rescheduled.

To lower the difficulty of writing suspend continuations, we now explicitly pass
that parcel to the continuation. At that point its stack has already been
checkpointed and it has been swapped out, so it is safe to treat it as a
suspended thread.


Former-commit-id: 2c5b550fd9b1da94696ae55807773a785f36d68b