Update the scheduler_suspend interface to allow blocking, and use a designated nop parcel.
This reduces the allocator churn for nop-ing. Whenever we need a nop we can just
transfer to the same thread that does nothing but call _schedule() in a loop. In
the future we could use this thread to perform maintenance tasks or other
intelligent work, or even just to sleep for a while.


Former-commit-id: b537658610701bb005b407ac07ee4984cb45d331