Remove special handling for tasks.
Tasks will not block in their execution, but may access thread data (e.g.,
hpx_thread_current_target()) and could continue. We have been trying to optimize
their execution by reusing the system stack to run them. This avoids stack
allocation for executing a task. Now that we have stack caching, this is no
longer a concern.

Given the added complexity of this path I have decided to remove the special
case handling. It reduces both the real lines-of-code involved in scheduling, as
well as the complexity that must be understood to understand the scheduler.


Former-commit-id: 31925f454c5e0be4cf7a891abca672e5dbdc63eb