Only open the hugetlbfs once at startup.
Previously, due to a misunderstanding about what hugetlbfs_unlinked_fd()
actually does, we opened a new file every time we needed to mmap() network
addressable memory. For applications that do lost of large allocations, e.g.,
sendrecv, this overflows the number of file descriptors we can open.

This patch simply opens the file descriptor once, in a constructor. This
resolves regression failures for photon+hugetlbfs.


Former-commit-id: 74e557ec96e11ccec22978180ccf115598e624b8