"Borrow" the stack when running an interrupt.
Interrupts can acquire and release LCO locks, and we need to keep track of
this. We do so on the lightweight thread stack, so the interrupt needs a stack
pointer. This is okay, since we are not allowing nested monitor calls so this
field is unused.

This reverts functionality from commit
6d3681b8e0328541e68a5156fe0f5b266d8e1560 [formerly dac0c4b6a598fa3fc2c83c2881ddf198766eb79e]. This used the worker structure to
track LCO depth, however this seems to be running into a gcc optimization
problem triggering a race. It's not clear if this was gcc's fault or our
fault---no other compilers generate the issue but that's not definitive.


Former-commit-id: ff0e3c7ad528ffa7e81aa4a959c9664aa3261a30