Only add -Wl,-rpath,${libdir} for static linking.
We assume that users linking to libhpx.so have libhpx.so in their library search
path already, however users linking to libhpx.a do not know that we link
jemalloc dynamically so they may not have the right search paths. This makes
sure libjemalloc.so can be found at runtime for static links.


Former-commit-id: 3bd5234f96ef577a3d185fa2c01f5f7cd24ffe3d