Improve instrumentation system's handling of time
A previous commit added a "beginning of time" - now the instrumentation system
uses this as the start time for each log. This way, user code can be relative
to the same start time and do so with less code than trying to expose
parts of the log tables to users.


Former-commit-id: de024865e70d10fce5a4439e2b670fde171e24e8