Restructure _schedule() interface---no functionality changes.
Traditionally, our scheduler worked in two phases.

1) A lightweight thread would call _schedele() to select a parcel to transfer
   to.

2) The lightweight thread called _transfer(), passing in the parcel that was
   returned from _schedule() along with a continuation that described what to do
   with the current thread after the transfer completed.

Recently, we moved towards using a unified scheduler_suspend() interface, that
would call _schedule() and _transfer() internally. This reduced the number of
lines of code in worker.c, but scheduler_suspend() is an external interface and
difficult to trace in a call-site specific context.

This patch unifies _schedule() and _transfer() into a single external
operation. _schedule() still selects a parcel to transfer to, but now simply
calls _transfer() directly. This architecture makes sense when we look at
_schedule() as a policy layer on top of _transfer(), further reduces the number
of lines of code in worker.c, and simplifies event tracing as scheduler_suspend
is now external-only.

Furthermore, this patch eliminates the return error code from
thread_transfer(). This code was never used (thus complicating the interface for
no reason), could not be recovered from even if it was used (the error is
delivered to the thread we're transferring to...what good is that?), and required
lots of useless `return HPX_SUCCESS;` and `dbg_check()` statements.


Former-commit-id: fbc2c5a2b0561f2dede177787e8d9c832d96c74b