Revert include error introduced in 934fdbf.
For whatever reason the worker structure has to always be available in the
parcel.h header because it is used in instrumentation declared static inline. I
disagree with this design---where possible instrumentation should be defined
static in the translation unit of its---but given the current structure we need
worker.h included in parcel.h unconditionally.


Former-commit-id: 34c9f42baff90a20a09ffb8debed3875a59028e8