Extend the HPX interface with hpx_lco_set_lsync and hpx_lco_set_rsync.
It makes sense to provide these semantics to programs that want it. Currently we
implement them in the obvious way, by allocating a future and forwarding to the
base hpx_lco_set() operation. We can potentially optimize this by using
scheduler_suspend(), since the future doesn't escape the thread that allocates
it.


Former-commit-id: 85c3985ca1297c7a219fa52956be5bd780629828