Added compound hpx_lco_wait_reset and hpx_lco_get_reset operations.
This functionality can be useful, and this compound atomic operation reduces the
number of locks, and-or messages, required. It is implemented in both local and
rendevous-remote form. Also upgraded remote lco_wait operations to use
scheduler_suspend.

Currently remote get/wait operations are not correctly handling errors. I have
added a parcel_launch_error operation to deliver an error instead of a value,
but it is unimplemented. Implementing it is complicated enough that it deserves
its own issue.


Former-commit-id: c922b33a6163fc2f678ccde795dc2a0e74c2a05b