Implement steal-last scheduler optimization.
This scheduler optimization makes a worker remember the victim that it last
stole from successfully, and tries to steal from it again. If no work is found,
it goes back to random stealing until a worker is found that it is successfully
able to steal from.

With stealing one parcel at a time, this doesn't seem to have any noticeable
difference in performance. For a carefully engineered use-case in parbench, where
an action affinitized to a core generates work periodically, this seems to marginally
reduce the number of failed steals. We expect such a scenario to be more common in
the distributed case where a large chunk of non-local work periodically shows up at
the network thread.


Former-commit-id: 081ce4ea9a3fcbbe068afe095a6bf988c259e291