Fix HPX-674.
If we cannot figure out the cpuset of a thread, try to get the
available CPUS from the job (using a resource-manager specific
environment variable such as ALPS_APP_DEPTH or SLURMS_JOB_CPUS_PER_NODE).

If none of those are set, try to use all of the available CPUS in
the system. If --hpx-threads is not set, we use this to determine
how many threads to run. Otherwise, we spawn as many threads as
the user wants on the number of CPUS we detected.


Former-commit-id: 0a7d16d4fa5ce1c126b970006d86ee01c4a57f4a