Extend the internal apps config with a HPX_APPS_LDFLAGS variable.
When we build .so contrib files like jemalloc.so, we need to make sure that
internal apps can find them. We tell libtool about the correct rpaths for this,
but libtool *does not* correctly propagate these rpaths during app link.

This patch fixes this issue by explicitly adding the rpaths we need for internal
app linking.