Append MPI_CFLAGS to HPX_APPS_CFLAGS.
The collbench performance test wants to be able to `#include <mpi.h>` based on
`HAVE_MPI`. We haven't ever exposed any `CFLAGS` for MPI to applications
though. This patch pushes the `MPI_CFLAGS` into the `HPX_APPS_CFLAGS` variable
to make this possible. We still aren't going to expose a path to MPI for
external applications through `hpx.pc`.