Don't let dlmalloc mmap() more memory.
We are only using dlmalloc to manage the heap and registered memory, in two
large segments, using mspaces. Unfortunately, dlmalloc will mmap() if it runs
out of memory in an mspace, so we turn that off. This is necessary for pwc since
we need to make sure that memory is registered.