Switch how we handle purging in jemalloc 4.0.
It used to be that we would just disable dirty page purging system-wide, because
this would conflict with network-registered mappings. In jemalloc 4.0 there is a
more complicated "cascade" from decommit to purge to dalloc, but it appears that
disabling purging prevents jemalloc from ever dallocing. This causes memory
exhaustion in a number of benchmarks.

This patch disables the arena-wide purge control we used, and instead installs a
no-op purge callback, which allows the cascade to pass to dalloc and thus munmap.