Fix our dlmalloc allocation.
This deals with the out-of-registered-memory issues that we have been
having. Basically, it uses system malloc to deal with all malloc/free calls (so
LD_PRELOAD something nice), and then uses an "anonymous" mspace to service
registered allocations. It sets dlmalloc's MMAP and MUNMAP redirects to a custom
mmap/munmap that we provide that forwards to the system's versions.

This allows the registered memory to grow as necessary. It has to set some other
dlmalloc variables to prevent issues, those can be seen in Makefile.am
CPPFLAGS. It can't detect out-of-memory for global memory---this will just
mmap() more space---but it will crash when we try and get offsets for that
memory.