Add an experimental process-based interface to an allreduce collective.
This change leaves the LCO allreduce intact as it used to be, but begins the
process of defining a process-level interface for collectives. The primary
difference is that the process collectives will have a different performance
expectation.

In particular, the point of the process collectives is that they have the
expectation that there will be a large number of participants that are broadly
distributed around the system, and should have efficient and scalable
implementations under such circumstances.

This abstraction is necessary for many HPC applications, and will be a
scalability bottleneck if they aren't implemented well. The interface is
currently considered experimental and will undergo changes prior to the 2.0
release.