Add a new reduction implementation for the process allreduce.
The previous tree reduction was designed to minimize contention, as well as to
retain all of the continuation information that we would need to broadcast the
reduced value. Now that we will be using a distributed continuation table, we no
longer need to store that continuation information.

This makes a simpler continuation structure more appealing, where we only keep
partially evaluated results around as opposed to all of the original data. This
"flat" reduce implementation is designed so that each worker thread has a
private slot to perform partial reductions into. It currently keeps track of the
number of remaining inputs using a fetch-and-add, but we can use a more scalable
counter in the future.