Removed previous tree-based local allreduce.
The previous tree-based version is useful, and should serve as an interesting
implementation, but was designed with the goal of retaining all of the incoming
parcels so that I could reuse them for continuation purposes.

Now that we have a distributed broadcast phase, I don't need this functionality
anymore. The flat allreduce performs better and is much simpler.