Avoid extra parcel copy in hpx_gas_map().
We conservatively cloned the parcel passed as the payload of
a marshalled action in the hpx_gas_map() implementation. This
change avoid the extraneous copy and replaces it by marking the
internal parcel as "nested" and the outer one as "retained". This
functionality requires that the parcel be the first member of the
payload. To meet this requirement, we pass the output stride in the
second parcel parameter (parcel->next) which is ignored in this
case.