More hpx-map reorganization.
Rename hpx_map_with_continuation to hpx_gas_bcast_with_continuation.
Rename hpx_map to hpx_gas_bcast.

Both of these above operations broadcast an action to the elements
of a global array.

In addition to these, hpx_gas_map performs a map operation---in the
traditional functional programming style---where it broadcasts an
action to the elements of a global array and stores the output in
another global array.