Generate a "core-map" during topology discovery.
When we run with hyper-threading enabled, multiple hardware threads are
mapped onto the same core. The core-map maintains a "cpu or hw thread id
to core id" mapping. This would let us have a "core affinity policy" where
we can pin the worker thread to a core (as opposed to a hw thread / cpu).