Add worker thread affinity policies for HPX.
The supported policies, as listed in help, are: default, hwthread, core, numa and none.
These determine how worker threads are pinned to PEs. At the time of topology discovery,
we generate cpusets for each resource (determined according to the global thread affinity
policy). After a worker thread starts, it binds itself according to this generated cpuset.