Add different (work-stealing) scheduling policies to HPX.
The work-stealing policy can now be chosen at runtime. The default is to
steal from a random victim. The "last" policy steals from the last successful
victim. In addition to these two already existing policies, a new policy "half"
has been added.

For now, the "half" policy asynchronously steals half of the work items from a
random victim. In the future, we might want to make it smarter--or perhaps the
default when stealing from a victim that lives in a different NUMA domain.

If a "victim" finds multiple "steal-half" requests in its work queue, it purges
the rest of the requests to avoid "recursive halving" of its work. Presently, a
random steal-half-threshold of 6 work items has been arbitrarily chosen.