Remove hpx_gas_map.
The `hpx_gas_map` interface was introduced to make it easier to operate on global arrays.
The point was to map an action on a global array and store the output in-place or a
different part of the global address space. The common-case operation (that of storing output
in-place) could also be achieved by simply mutating the global data. The whole interface
turned out to be a little awkward and should be revisited at some point.