action: Support variadic number of marshalled arguments.
Oftentimes, we want to call an action with a variable number of
marshalled buffers. The usual way to do this is to wrap the arguments
in a struct and serialize it as a contiguous buffer. This is cumbersome.
We want to be able to register a marshalled action with multiple
(pointer, size) marshalled buffers. This is achieved by extending the
action table to remember the number of marshalled arguments the action
was registered with.