Extend the GAS interface to support GAS attributes.
We need to be able to specify attributes for GAS allocations. This change
extends `hpx_gas_alloc` and `hpx_gas_calloc` with an attribute. Further,
a set of GAS allocation functions suffixed with `_attr` are also added.
The attributes currently defined are: none, LB (load-balancing) and RO
(read-only). The PGAS implementation presently ignores attributes.